Include the animation's ID in the metadata
This is useful on the front end. The UI doesn't want to load all of the
animations that might exist into memory (including their frame data), so
the server returns a list of metadata objects. This contains everything
except for the ID and the frame data.

Well, that's a problem, because we need to be able to go from the
animation metadata back to the to the animation in UI for the purposes
of editing it.

This change also re-does how animation IDs are handled on save. Since we
might be assigning a new ID (because it's a create), and the Animation
object that's passed in from the server is immutable, we need some other
way to pass the newly assigned ID along to save it in the database.