Blow up backwards compatibility and move to a mutlicast world
This change makes non-backwards compatible changes. It removes the IP
information, assumes everything is multicast, and renamed "dmx_offset"
to "channel_offset" to acknowledge that we're now in a e1.31 centric
world.

It also totally removes the motors from the database. I never actually
used them, and the source of truth is now the creature's config file.
I might have the console be able to read the creature's config file and
adjust things in the UI from there. It did not make sense to have what
looked like two sources of truth when there actually wasn't.